requirements.txt
pymodbus==3.8.6
pyserial>=3.5
